change testcase according to feedback
