+++ src/test/java/spoon/test/executable/ExecutableTest.java
@@ -35,6 +35,9 @@ import spoon.testing.utils.ModelUtils;
 
 import java.util.List;
 
+import static org.hamcrest.CoreMatchers.anyOf;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -125,5 +128,21 @@ public class ExecutableTest {
 		assertEquals(valueOf, shadowValueOf);
 		assertEquals(valueOf.getDeclaration(), shadowValueOf.getDeclaration());
 		new ContractVerifier(shadowValueOf.getParent(CtPackage.class)).checkShadow();
+	}
+
+	@Test
+	public void testGetAllExecutablesOnTypeImplementingNestedInterface() {
+		// contract: implicit static nested interfaces are correct handled in getAllExecutables.
+		Launcher launcher = new Launcher();
+		launcher.addInputResource("src/test/resources/extendsStaticInnerType");
+		CtModel model = launcher.buildModel();
+		CtType<?> type = model.getAllTypes().stream().filter(v -> v.getSimpleName().contains("BarBaz")).findAny().get();
+		int expectedNumExecutablesInJDK8 = 13;
+		int expectedNumExecutablesPostJDK8 = 14;
+		int numExecutables = type.getAllExecutables().size();
+		assertThat(numExecutables, anyOf(
+				equalTo(expectedNumExecutablesInJDK8),
+				equalTo(expectedNumExecutablesPostJDK8))
+		);	
 	}
 }
